Refactor: change ID to long

+++ src/main/java/mhjohans/userapi/model/User.java
@@ -4,24 +4,26 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "users")
-public record User(@Id int id, String name, String username, String email, Address address, String phone, String website, Company company) {
+public record User(@Id long id, String name, String username, String email, Address address, String phone,
+                   String website, Company company) {
 
     /**
      * Create a new User from an existing User and a custom ID
-     * 
+     *
      * @param user existing User object
-     * @param id custom ID
-     * 
+     * @param id   custom ID as long value
      * @return a new User with the custom ID
      */
-    public User(User user, int id) {
+    public User(User user, long id) {
         this(id, user.name(), user.username(), user.email(), user.address(), user.phone(), user.website(), user.company());
     }
-    
+
     public record Address(String street, String suite, String city, String zipcode, Geo geo) {
-        public record Geo(String lat, String lng) {}
+        public record Geo(String lat, String lng) {
+        }
     }
 
-    public record Company(String name, String catchPhrase, String bs) {}
+    public record Company(String name, String catchPhrase, String bs) {
+    }
 
 }
